[checker][refactor] Move constraint solver into type system module
